Set up layout for python files similar to the old gui
